+++ cogs/__init__.py
@@ -4,6 +4,7 @@ EXTENSIONS = (
     "cogs.general",
     "cogs.fun",
     "cogs.casino",
+    "cogs.economia",
     "cogs.games",
     "cogs.music",
     "cogs.lyrics",
